pyproject.toml
[tool.poetry]
name = "gmail-tag"
version = "0.1.0"
description = ""
authors = ["diegoasua <baxuaa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-auth = "^2.29.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.131.0"
openai = "^1.30.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
